ShinyPubMedVectorSearch/requirements.txt
shiny~=1.1.0
psycopg~=3.2.2
pgvector~=0.3.3
python-dotenv~=1.0.1
sentence_transformers~=3.1.1
